标签分页查询

+++ blog-common/src/main/java/top/ddgotxdy/common/model/PageResult.java
@@ -14,7 +14,7 @@ import java.util.List;
 @Data
 public class PageResult<T> {
     @ApiModelProperty("总页数")
-    private Integer totalNumber;
+    private Long totalNumber;
     @ApiModelProperty("返回结果")
     private List<T> data;
 }
